textSubmitter widget update WIP
